mark conc test as big
